Task 1
Add a new API call PATCH constructionStages/{id} which to allow the API users to edit an arbitrary field of a selected (by id) construction stage. The API should touch only the fields which are sent by the user. Add validation which to ensure that if status field is sent it is either NEW, PLANNED or DELETED and throw a proper error if it is not.

Add another DELETE constructionStages/{id} API call which changes the status of the selected resource to DELETED.