Task 2
Write a validation system which checks every posted field against a set of rules as follows:

name is maximum of 255 characters in length
start_date is a valid date&time in iso8601 format i.e. 2022-12-31T14:59:00Z
end_date is either null or a valid datetime which is later than the start_date
duration is skipped because it should be automatically calculated based on start_date, end_date and durationUnit
durationUnit is one of HOURS, DAYS, WEEKS or can be skipped (which fallbacks to default value of DAYS)
color is either null or a valid HEX color i.e. #FF0000
externalId is null or any string up to 255 characters in length
status is one of NEW, PLANNED or DELETED and the default value is NEW.
You should throw proper errors if a rule is not met.