Task 3
Set a logic which automatically calculates duration based on start_date, end_date and durationUnit as you know that:

start_date is required and is a valid date&time in iso8601 format i.e. 2022-12-31T14:59:00Z
end_date is either null (then duration is also null) or a valid datetime which is later than the start_date
durationUnit is one of HOURS, DAYS, WEEKS where DAYS is the default fallback.
duration is a positive float value calculated in precision of whole hours (ignore minutes and seconds if any)
a week has 7 days and one day has 24 hours